pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "news-pipeline"
version = "0.1.0"
description = "Modular news processing pipeline"
authors = [
    {name = "Euan", email = "euan.m.campbell@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    # Shared dependencies
    "boto3>=1.42,<2.0",
    "pyyaml>=6.0.3,<7.0",
    "python-dotenv>=1.2,<2.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Ingest service dependencies
[tool.poetry.group.ingest.dependencies]
psycopg2-binary = ">=2.9.11,<3.0"
feedparser = ">=6.0.12,<7.0"
requests = ">=2.32,<3.0"
python-dateutil = ">=2.9,<3.0"
trafilatura = ">=2.0,<3.0"
readability-lxml = ">=0.8,<1.0"

# Normalize service dependencies
[tool.poetry.group.normalize.dependencies]
spacy = ">=3.7"
spacy-transformers = ">=1.3"
pyarrow = ">=18.0"
sentence-transformers = ">=2.2"
numpy = ">=1.26"
pycountry = ">=24.6"
geopy = ">=2.4"

# Cluster service dependencies
[tool.poetry.group.cluster.dependencies]
hdbscan = ">=0.8.33"
numpy = ">=1.26"
pyarrow = ">=18.0"

# API service dependencies
[tool.poetry.group.api.dependencies]
fastapi = ">=0.115"
uvicorn = {version = ">=0.34", extras = ["standard"]}
pydantic = ">=2.10"
pyarrow = ">=18.0"

# Development dependencies
[tool.poetry.group.dev.dependencies]
pytest = ">=9.0"
pandas = ">=2.0"

[tool.pytest.ini_options]
testpaths = [
    "services/news-ingest/tests",
    "services/news-normalize/tests",
    "services/news-cluster/tests",
    "services/news-api/tests",
]
pythonpath = [
    "services/news-ingest",
    "services/news-normalize",
    "services/news-cluster",
    "services/news-api",
]
